feat(g2p): implement clickable phoneme and variant selection components
- Introduced `ClickablePhoneme` component for interactive phoneme display, enhancing user engagement with phonetic elements.
- Added `IpaSequence` component to manage syllable display and phoneme interactions, improving the structure of the transcription display.
- Implemented `VariantSelector` for selecting different phonetic variants, streamlining the user experience when dealing with multiple pronunciations.

These changes enhance the functionality and usability of the transcription display in the G2P component.
